Change slave address type;
The I2C message (SlaveToMasterMessage) treats address as a single byte
value. The address is also stored in the EEPROM as a single byte. Thus
it makes little sense to have the address as integer in the code.

+++ arduino/slave/slave.h
@@ -84,7 +84,7 @@ private:
   ChangeHandler handler;
   Adafruit_NeoPixel* leds[3];
   
-  volatile int address;
+  volatile uint8_t address;
 
   uint8_t switchStates;
   uint8_t previousSwitchStates;
